レビュー箇所の修正

+++ app/controllers/items_controller.rb
@@ -40,10 +40,11 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-
     item = Item.find(params[:id])
-    if item.destroy
-      redirect_to root_path
+    if item.user_id == current_user.id
+        if item.destroy
+          redirect_to root_path
+        end
     end
   end
